@@ -1,30 +1,32 @@
+type CharacterStatus = "Dead" | "Alive" | "unknown" | "info";
+
 export type Character = {
-  id: number
-  name: string
-  status: string
-  species: string
-  type: string
-  gender: string
+  id: number;
+  name: string;
+  status: CharacterStatus;
+  species: string;
+  type: string;
+  gender: string;
   origin: {
-    name: string
-    url: string
-  }
+    name: string;
+    url: string;
+  };
   location: {
-    name: string
-    url: string
-  }
-  image: string
-  episode: string[]
-  url: string
-  created: string
-}
+    name: string;
+    url: string;
+  };
+  image: string;
+  episode: string[];
+  url: string;
+  created: string;
+};
 
 export type Characters = {
   info: {
-    count: number
-    pages: number
-    next: string | null
-    prev: string | null
-  }
-  results: Character[]
-}
+    count: number;
+    pages: number;
+    next: string | null;
+    prev: string | null;
+  };
+  results: Character[];
+};
