@@ -1,19 +1,19 @@
 export type Episode = {
-  id: number
-  name: string
-  air_date: string
-  episode: string
-  characters: string[]
-  url: string
-  created: string
-}
+  id: number;
+  name: string;
+  air_date: string;
+  episode: string;
+  characters: string[];
+  url: string;
+  created: string;
+};
 
 export type Episodes = {
   info: {
-    count: number
-    pages: number
-    next: string | null
-    prev: string | null
-  }
-  results: Episode[]
-}
+    count: number;
+    pages: number;
+    next: string | null;
+    prev: string | null;
+  };
+  results: Episode[];
+};
