@@ -1,3 +1,3 @@
-export * from "./episodes"
-export * from "./characters"
-export * from "./locations"
+export * from "./episodes";
+export * from "./characters";
+export * from "./locations";
